Fix missing headers
