+++ src/extensions/gl_ext_OES_point_sprite.c
@@ -1,0 +1,5 @@
+#include "gl_ext_common.h"
+#include <GLES/glext.h>
+EXT_REGISTER("GL_OES_point_sprite")
+int ext_link_dummy_OES_point_sprite;
+/* Enable handled in gl_api_state.c, coord replace in gl_api_texture.c */
